complete implementation of usecase

+++ src/main/java/ir/store/java/object/feature/impl/ElectricalVehicleDAOImpl.java
@@ -7,12 +7,12 @@ import ir.store.java.object.model.ElectricalVehicle;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 
 public class ElectricalVehicleDAOImpl implements ElectricalVehicleDAO {
-
+    private static DataSource dataSource=new DataSource();
     @Override
     public List<ElectricalVehicle> getAllElectricalVehicle() {
-        DataSource dataSource=new DataSource();
         Connection con=dataSource.createConnection();
         Statement statement=null;
         ResultSet rs=null;
@@ -46,33 +46,39 @@ public class ElectricalVehicleDAOImpl implements ElectricalVehicleDAO {
     public void addElectricalVehicle(ElectricalVehicle electricalVehicle) {
         Connection dbConnection=null;
         PreparedStatement statement=null;
-        String sql="insert into electrical_vehicle values(?,?,?,?,?,?)";
-        try
-        {
-            DataSource dataSource=new DataSource();
-            dbConnection=dataSource.createConnection();
-            statement=dbConnection.prepareStatement(sql);
-            statement.setInt(1,electricalVehicle.getId());
-            statement.setString(2,electricalVehicle.getName());
-            statement.setString(3,electricalVehicle.getType());
-            statement.setInt(4,electricalVehicle.getPower());
-            statement.setInt(5,electricalVehicle.getStock());
-            statement.setDouble(6,electricalVehicle.getPrice());
+        if(!existCheck(electricalVehicle.getId())) {
+            String sql = "insert into electrical_vehicle values(?,?,?,?,?,?)";
+            try {
+                DataSource dataSource = new DataSource();
+                dbConnection = dataSource.createConnection();
+                statement = dbConnection.prepareStatement(sql);
+                statement.setInt(1, electricalVehicle.getId());
+                statement.setString(2, electricalVehicle.getName());
+                statement.setString(3, electricalVehicle.getType());
+                statement.setInt(4, electricalVehicle.getPower());
+                statement.setInt(5, electricalVehicle.getStock());
+                statement.setDouble(6, electricalVehicle.getPrice());
 
-            statement.executeUpdate(sql);
-            System.out.println("Record is inserted into electricalVehicle table for Good: "+electricalVehicle.getName());
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        finally {
-            finish(dbConnection, statement);
+                statement.executeUpdate(sql);
+                System.out.println("Record is inserted into electricalVehicle table for Good: " + electricalVehicle.getName());
+
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                finish(dbConnection, statement);
+            }
+        }else {
+            System.out.println("there is a shoe with id="+electricalVehicle.getId());
+            System.out.print("Would you like to update this shoe? \n\t1.Yes \n\t2.No \n respone:");
+            if(new Scanner(System.in).nextInt()==1) updateElectricalVehicle(electricalVehicle);
+
         }
 
     }
 
     @Override
     public ElectricalVehicle getElectricalVehicle(int electricalVehicleId) {
-        DataSource dataSource=new DataSource();
+
         Connection connection=dataSource.createConnection();
         Statement stmt=null;
         ResultSet rs=null;
@@ -80,16 +86,16 @@ public class ElectricalVehicleDAOImpl implements ElectricalVehicleDAO {
             String query="select * from electrical_vehicle where  id="+electricalVehicleId;
             stmt=connection.createStatement();
             rs=stmt.executeQuery(query);
-
-                ElectricalVehicle good=new ElectricalVehicle();
-                good.setId(rs.getInt("Id"));
-                good.setName(rs.getString("name"));
-                good.setPrice(rs.getInt("price"));
-                good.setStock(rs.getInt("stock"));
-                good.setPower(rs.getInt("power"));
-                good.setType(rs.getString("type"));
-
-                return good;
+            // exploring
+            if(rs.wasNull()) return  null;
+            ElectricalVehicle good=new ElectricalVehicle();
+            good.setId(rs.getInt("Id"));
+            good.setName(rs.getString("name"));
+            good.setPrice(rs.getInt("price"));
+            good.setStock(rs.getInt("stock"));
+            good.setPower(rs.getInt("power"));
+            good.setType(rs.getString("type"));
+            return good;
 
 
         } catch (SQLException e) {
@@ -119,25 +125,35 @@ public class ElectricalVehicleDAOImpl implements ElectricalVehicleDAO {
 
     @Override
     public void updateElectricalVehicle(ElectricalVehicle good) {
-        Connection dbConnection=null;
-        Statement statement=null;
-        String sql="update electrical_vehicle set name= "+good.getName()+" where id ="+good.getId();
-        try{
-            DataSource dataSource=new DataSource();
-            dbConnection=dataSource.createConnection();
-            statement=dbConnection.prepareStatement(sql);
-            statement.executeUpdate(sql);
-            System.out.println("Record is updated into electrical_vehicle table for id :"
-                    +good.getId());
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        finally {
-            finish(dbConnection, statement);
+        Connection connection=null;
+        PreparedStatement statement=null;
+        if(existCheck(good.getId())) {
+            String query = "update electrical_vehicle set name=?,set type=? ,set power=?" +
+                    ",set stock=? , set price=? where id =" + good.getId();
+            try {
+                connection = dataSource.createConnection();
+                statement = connection.prepareStatement(query);
+                statement.setString(1, good.getName());
+                statement.setString(2, good.getType());
+                statement.setInt(3, good.getPower());
+                statement.setInt(4, good.getStock());
+                statement.setDouble(5, good.getPrice());
+                statement.execute();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                finish(connection, statement);
+            }
+        }else {
+            System.out.println("There aren't any electrical vehicle with id= " + good.getId());
+            System.out.print("Would you like to add this reading case? \n\t1.Yes \n\t2.No \n respone:");
+            if (new Scanner(System.in).nextInt() == 1) {
+                updateElectricalVehicle(good);
+            }
         }
     }
 
-    private void finish(Connection dbConnection, Statement statement) {
+    private void finish(Connection connection, Statement statement) {
         if(statement!=null){
             try {
                 statement.close();
@@ -145,11 +161,11 @@ public class ElectricalVehicleDAOImpl implements ElectricalVehicleDAO {
                 e.printStackTrace();
             }
         }
-        if(dbConnection!=null)
+        if(connection!=null)
         {
             try
             {
-                dbConnection.close();
+                connection.close();
             } catch (SQLException e) {
                 e.printStackTrace();
             }
@@ -160,21 +176,25 @@ public class ElectricalVehicleDAOImpl implements ElectricalVehicleDAO {
     public void deleteElectricalVehicle(int electricalVehicleId) {
         Connection dbConnection=null;
         Statement statement=null;
-        String sql="delete from electrical_vehicle where Id="+electricalVehicleId;
-        try{
-            DataSource dataSource=new DataSource();
-            dbConnection=dataSource.createConnection();
-            statement=dbConnection.prepareStatement(sql);
-            statement.executeUpdate(sql);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        finally {
-            finish(dbConnection,statement);
-        }
+        if(existCheck(electricalVehicleId)) {
+            String sql = "delete from electrical_vehicle where Id=" + electricalVehicleId;
+            try {
+                DataSource dataSource = new DataSource();
+                dbConnection = dataSource.createConnection();
+                statement = dbConnection.prepareStatement(sql);
+                statement.executeUpdate(sql);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                finish(dbConnection, statement);
+            }
+        }else System.out.println("There aren't any electrical vehicle with id= "+electricalVehicleId);
 
+
+}
+
+    private boolean existCheck(int id) {
+        boolean exist=(getElectricalVehicle(id)!=null);
+        return exist;
     }
-
-
-
 }
